Move homepage URL to a resValue in build config
We're removing it from the list of non translatable strings. Having
it in the build config allows us to specify different start pages
per flavour. In this case we're interested in providing a different
one for the Chinese market as some of the experiences in the start
page cannot be either viewed from mainland China or simply take a
lot to load.

So far we're using about:blank but we can change that in the future.