[HVR] Fix flavor detection when deciding about requesting SSID
We've recently added some code for HVR in China that was meant to
prevent Wolvic to call getSSID() because that's considered
recollection of personal data even if we only use it to show the
SSID in the UI.

The problem is that we were comparing the lowercase name of the
flavor with "mainlandChina" with capital C, so this was obviously
never matching. Correct that and also replace the displayed tooltip
text from "unknown" to "unavailable" which is more precise.